rename : inital_input에서 input으로 패키지 이름 변경
