+++ moview/service/interviewee_input/interviewee_input_service.py
@@ -1,7 +1,7 @@
 from typing import Any
 
-from moview.modules.initial_input.initial_input_analyzer import InitialInputAnalyzer
-from moview.modules.initial_input.initial_question_giver import InitialQuestionGiver, InitialQuestionParseError
+from moview.modules.input.input_analyzer import InputAnalyzer
+from moview.modules.input.initial_question_giver import InitialQuestionGiver, InitialQuestionParseError
 from moview.repository.interviewee_data_repository import IntervieweeDataRepository, MongoConfig
 from moview.repository.entity.interviewee_data_main_document import IntervieweeDataEntity
 from moview.repository.entity.interviewee_data_subdocument import *
@@ -13,7 +13,7 @@ class IntervieweeInputService:
         # todo MongoConfig 나중에 삭제하거나 수정할 필요 있을 듯.
         self.repository = IntervieweeDataRepository(mongo_config=MongoConfig())
 
-        self.initial_input_analyzer = InitialInputAnalyzer()
+        self.initial_input_analyzer = InputAnalyzer()
         self.initial_question_giver = InitialQuestionGiver()
 
     def ask_initial_question_to_interviewee(self, session_id, interviewee_name: str,
