+++ tests/modules/initial_input/test_inital_question_giver.py
@@ -1,6 +1,6 @@
 import unittest
 from tests.common_code_for_test import is_not_none_string
-from moview.modules.initial_input.initial_question_giver import InitialQuestionGiver
+from moview.modules.input.initial_question_giver import InitialQuestionGiver
 
 
 class TestInitialQuestionGiver(unittest.TestCase):
